refactor(terminal): move is_interactive out of logger into terminal module (#176)
is_interactive is a TTY/terminal capability check, unrelated to logging.
Home it in a dependency-free terminal.py so logger, confirmer, and cli all
import it without inverting the logger->cli dependency direction. logger
re-exports it via its namespace, so existing test patch sites still resolve.
