+++ src/pcswitcher/logger.py
@@ -18,6 +18,7 @@ from rich.console import Console
 from rich.text import Text
 
 from pcswitcher.config import LogConfig
+from pcswitcher.terminal import is_interactive
 
 # Register custom FULL level (15) with stdlib logging
 FULL = 15
@@ -319,20 +320,6 @@ class WarningCaptureHandler(logging.Handler):
             self._sink.add_warning(format_log_line(record))
         except Exception:
             self.handleError(record)
-
-
-def is_interactive(console: Console) -> bool:
-    """Return True only when the run is fully interactive on both stdin and stdout.
-
-    A run is interactive only if BOTH ends are a terminal: a real terminal on
-    stdout (``console.is_terminal``) so a live UI / prompt is actually visible,
-    AND a TTY on stdin (``sys.stdin.isatty()``) so the user can actually answer
-    a prompt. Requiring both keeps logging setup and the confirmer in agreement
-    under mixed redirection (e.g. stdout is a TTY but stdin is ``/dev/null``):
-    a single split signal previously let the live UI + UILogHandler activate
-    while confirmations silently fell back to ``--allow-*`` flags.
-    """
-    return console.is_terminal and sys.stdin.isatty()
 
 
 def generate_log_filename(session_id: str) -> str:
